chore(Accounts): move navigation helpers to utils [YTFRONT-4896]
