+++ packages/ui/src/ui/pages/query-tracker/Navigation/NodeList/NodeListRow.tsx
@@ -5,8 +5,8 @@ import cn from 'bem-cn-lite';
 import {getIconNameForType} from '../../../../utils/navigation/path-editor';
 import AwesomeIcon from '../../../../components/Icon/Icon';
 import {Button, DropdownMenu, Icon, Text} from '@gravity-ui/uikit';
-import {isFolderNode} from '../helpers/isFolderNode';
-import {isTableNode} from '../helpers/isTableNode';
+import {isFolderNode} from '../../../../utils/navigation/isFolderNode';
+import {isTableNode} from '../../../../utils/navigation/isTableNode';
 import StarFillIcon from '@gravity-ui/icons/svgs/star-fill.svg';
 import StarIcon from '@gravity-ui/icons/svgs/star.svg';
 import TextIndentIcon from '@gravity-ui/icons/svgs/text-indent.svg';
