+++ packages/ui/src/ui/pages/query-tracker/module/queryNavigation/actions.ts
@@ -35,8 +35,8 @@ import {JSONParser} from '../api';
 import {toggleFavourite} from '../../../../store/actions/favourites';
 import {createNestedNS} from '../../../../../shared/utils/settings';
 import {NAMESPACES} from '../../../../../shared/constants/settings';
-import {isTableNode} from '../../Navigation/helpers/isTableNode';
-import {isFolderNode} from '../../Navigation/helpers/isFolderNode';
+import {isTableNode} from '../../../../utils/navigation/isTableNode';
+import {isFolderNode} from '../../../../utils/navigation/isFolderNode';
 import {loadTableAttributes} from '../../Navigation/api/loadTableAttributes';
 import {QueryEngine} from '../../../../../shared/constants/engines';
 import {createEmptyQuery, loadCliqueByCluster, updateQueryDraft} from '../query/actions';
