+++ packages/ui/src/ui/utils/navigation/isFolderNode.ts
