+++ packages/ui/src/ui/utils/navigation/isTableNode.ts
